requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-ldap==3.4.4
cryptography==41.0.7
bcrypt==4.1.2
WTForms==3.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
ipaddress==1.0.23
validators==0.22.0
